Backend/ecofood/requirements.txt
django==4.2.11
djangorestframework==3.14.0 
drf-yasg==1.21.7
mysqlclient==2.2.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1